+++ src/scatrans/tl/_common.py
@@ -16,7 +16,7 @@ try:
 
     VERSION = _version.version
 except (ImportError, AttributeError):
-    VERSION = "0.10.1"
+    VERSION = "0.10.2"
 
 logger = logging.getLogger(__name__)
 logger.addHandler(logging.NullHandler())